BackEnd/Python/FLASK_SQLALCHEMY/requirements.txt
python-dotenv==1.0.1
mysqlclient==2.1.1
SQLAlchemy==2.0.35
Flask==3.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
